Double 로 변경

+++ src/data/SaleInfo.java
@@ -2,16 +2,16 @@ package data;
 
 public class SaleInfo implements Comparable<SaleInfo> {
     private final String partner;
-    private final int quantity;
+    private final Double quantity;
 
-    public SaleInfo(String partner, int quantity) {
+    public SaleInfo(String partner, Double quantity) {
         this.partner = partner;
         this.quantity = quantity;
     }
 
     @Override
     public int compareTo(SaleInfo o) {
-        return Integer.compare(o.quantity, this.quantity);
+        return Double.compare(o.quantity, this.quantity);
     }
 
     @Override
@@ -26,7 +26,7 @@ public class SaleInfo implements Comparable<SaleInfo> {
         return partner;
     }
 
-    public int getQuantity() {
+    public Double getQuantity() {
         return quantity;
     }
